Workaround for Intel/Vulkan fail
It looks like there's an issue in here that shows up for Vulkan/Intel. The result is a client crash due to memory access violation. This seems to step from passing a zero-dimensioned array to the Buffer method. This throws an exception (which is caught) for 'no valid memory'.

Passing the resulting 'LinesVertexBuffer' into Command.SetVertexBuffer() then results in the memory access violation.

Arguably, the library should be smarter than this, but we'll defensively set the buffer to null and rely on the check in the drawing code to sidestep potential trouble.