FEM_3D_Project
This behaves very oddly with a lot of uninitialized arrays being used with default values that are not zero or otherwise understood. I've tried to replicate the default values observed under the debugger. I used MSVC2019 for comparison.

The original code under MSVC2019 produces different values to the reference file (presumably due to the use of uninitialized values in the arrays leading to different starting points). The C# and MSVC2019 builds produce numerically equivalent results, however.

This whole routine seems fishy - the lack of proper initialization is worrying. Using default values (0) causes a pivot fail in the original code (pivot is zero due to the lack of deltas).