Keyholer experiment
ILB in Variance was showing some keyholer results that I wasn't happy with. Inspection showed that rays were now being cast in the wrong direction (across the cutter rather than outward). This commit brings an empirical fix by moving the inversion from Y to X in the raycaster. This needs thorough testing....

Additionally, an option for allowing orthogonal rays to override min length non-ortho has been added, but not active yet. It gives good quality results, but need to mull whether I like the results or not compared to the previous behavior.

ILB16 looks a lot better with these changes, but again, testing is required to check nothing else broke.