Switch off edge extension in the keyholer
This causes issues for Quilt decomp, due to notching of otherwise-orthogonal geometry. Will need to see whether this is OK for all usage scenarios, otherwise an option may be needed.
This was originally added due to issues with early Clipper2 handling of coincident edges. The library has improved since that time, so this may no longer be needed.