Keyhole enhancements
This commit changes orthogonal insertion preference to only be effective when orthogonal geometry is present in the outers and cutters. The orthogonal preference exists to try and maintain contouring options; if cutters/outers are not orthogonal, this requirement is no longer effective. Removing this constraint then allows for better insertion locations to be identified in these cases due to the additional freedom.
Additionally, the keyholer now performs a reverse walk, in addition to the existing sequence. This reverse walk is added to throw rays in the opposite direction. By doing this, we catch minimums that were missed from the initial geometry sequence because the emission is order-sensitive.