Extend behavior of keyholer
It can be useful to flip the processing of the keyhole, e.g. in Quilt, without also triggering the full bidirectional processing. To that end, refactor so that reverse evaluation can be requested, and rename the reverse walk parameter to indicate its bidirectional nature.
This has solved the Quilt case where lines were being bisected. Review in Variance is pending.