Decomposition hackery
This is awkward. We rely on the integer values to squelch the small irregularities introduced into etches that are cut by the edge insertion. Floats preserve these notches and screw up the decomposition. So we have to map to a large integer grid and then size down before returning the geometry. This isn't ideal.