Keyholer : fudge to close gaps
Something changed and it appears a fudge factor is required to reliably close gaps injected by the keyholer.
I think this may be a glancing edge issue and this fudge factor is enough to force a resolution. I'm not entirely comfortable with the empirical nature of this, but testing shows it works.