Adjusting to use shape settings
Parameters in calls are not needed - the shape settings already owns a copy if the user has done things properly and constructed the settings before generating the shape.
For some cases, we need to capture some additional information (e.g. tip bias type), hence reduction of the tip computation to a minimal set of arguments and also two public methods now for processCorners (in case the user needs a variation on top of the shape-defined values). The gTipBias value needs to be reviewed with the user application - might be redundant as well.