Adjust approach to tracking cage computation, error handling
Forcing the cage computation in the constructor prevented deferred cage computation because tips, etc. were already set and skipped.
Instead, track if cage is computed and, if not, compute it when the corners are processed (which is the last moment possible).
Also expose a string for last error, which allows for unit testing for custom shape.