Fix enclosure query code and revised test to match results
If two completely disconnected geometric entities were queried for enclosure, the old approach would yield the wrong result. Resolve this by performing a basic area check. If the same area post-union is computed as for the incoming polygons, no overlap can have existed (the union essentially subtracting the overlap area from the incoming total area).
This fixes the test results to be as-expected, so the test has been revised.