Major bug fix / rework
So, the design for 'box' was flawed due to basing it on a problematic GeoLibRectangle design. GeoLibRectangle had its X and Y at the center and then used width/height to compute the left/right, bottom/top values. This was intended to make it easier to apply rotations/mirroring, but turns out to be a really bad idea because the integer grid of layout cannot deal with this for odd-value dimensions or locations (e.g. 1 unit wide or tall, leading to a 0.5 value that cannot be located on the grid). Bad developer, no cookie.
This problem showed up in the GDS output for one of the decomposition tests, where one rectangle was zero-height in GDS, but fine in OASIS and closer inspection revealed other odd mismatches.
To fix this, GeoLibRectangle was reworked to have the lower left corner has the XY and then width and height are simply added to get right and top.

OASIS reader was patching this discrepancy and this was now removed because the OASIS design actually makes sense in the light of this realization.

Unit test coverage suggests all is fine after this change.