Cell and minimum/maximum changes
Cell takes a following change from the previous commit for the rectangle alteration.

For our GDS writer, boxes are written as boundaries (polygons) to match other tool behavior (e.g. KLayout). To allow boxes to be recognized on-load, new code is added that tries to identify a box by checking for the number of points, the equivalence of all height and width and then that angles are 90 degrees only. If this is matched, a box is added, otherwise a polygon will be added.

I also noticed that the minimum/maximum queries were essentially doing nothing because the supplied point wasn't getting adjusted. A reference needed to be passed in all cases. This has now been corrected.