WIP for repetition handling
This is fairly invasive. The existing code 'bakes' in explicit positioning information when loading layout and explicit positioning of instances is set. Moving to a repetition system (inspired by gtsdk) should eventually allow us to be more intelligent here, allowing for more efficient OASIS exports.
This particular commit brings the repetition class and this is still WIP. The aim is that this works seamlessly for the existing tests to make sure we read/write the data properly before breaking new ground elsewhere.
The naive pitch from before is replaced by row and column pitches (mapping to the rowVector and colVector internally), resulting from the desire to handle irregular arrays.
Setting up explicit arrays is next on the table.