Massive update to try and get GTK and Veldrid playing nicely.
This moves the source build for Veldrid because it's too difficult to manage the native library; the nuget has it anyway.
TechPizzaDev found a way to make GTK not explode with modern GTK which has some legitimate threading behavioral constraints.
Discarded the Mac implementations as I have no hardware and no user interest. Windows is dropped from the solution for now, but will return (WPF) eventually.
Note that the test in this commit is the spinning rainbow square until I sort the mess out.