Implicit needs a different approach
The deferred nature of the names means we cannot rely on the previous approach - the cell might not exist yet. Instead move to an array of cell data and using that, we track each cell's name to forcibly apply in a post-process. We also have the potential for similar shenanigans with text, so track the text data for each cell and apply that the same way. This might seem like overengineering.
This is untested so far, but prompted by failed tests for the previous approach. I'm hoping this will work out.
I've implemented with Parallel.For with the option to go to a basic for loop if needed.

I have some concerns that cellIndex might get out of step with the text handling due to the cellIndex being bumped in cellname_implicit, but we'll see if that is a possibility or not.