Sorting out CBLOCK test coverage
Also found a weird additional element being added in some cases. Added a hack that still passes for all existing tests, but I'm not confident in it.
Replaced the f_rep3.oas with a KLayout version for testing, noting that gdstk is not bug-free and might be generating an odd file.